+++ src/resources/quote.ts
@@ -632,6 +632,9 @@ export namespace _quote {
     id?: string;
   }
   export interface subscription_create_sub_for_customer_quote_params {
+    po_number?: string;
+  }
+  export interface subscription_create_sub_for_customer_quote_params {
     plan_id: string;
   }
   export interface subscription_create_sub_for_customer_quote_params {
@@ -1425,6 +1428,9 @@ export namespace _quote {
   }
   export interface subscription_create_sub_items_for_customer_quote_params {
     id?: string;
+  }
+  export interface subscription_create_sub_items_for_customer_quote_params {
+    po_number?: string;
   }
   export interface subscription_create_sub_items_for_customer_quote_params {
     trial_end?: number;